Reduce variations of translate command because of ambiguity
It's difficult to support 0 and 1 specified language because there's no
easy way to tell where the language arguments end and where the text to
translate begins.

Limit the variations:

for normal translation, always specify the "from" and "to" languages. If
the user wants the "from" language to be detected, they can provide a
'?'

for a reply translation, either no languages can be specified in which
case "from" is detected and "to" is default langauge (english), or both
"from" and "to" have to be specified (from can be '?' if it has to be
detected)